+++ PinkSea.Gateway/Program.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using PinkSea.Gateway.Endpoints;
 using PinkSea.Gateway.Models;
 using PinkSea.Gateway.Services;
 
@@ -19,71 +20,10 @@ builder.Services.AddHttpClient(
 
 var app = builder.Build();
 app.UseStaticFiles();
-app.MapGet(
-    "/{did}/oekaki/{rkey}", 
-    async ([FromRoute] string did, [FromRoute] string rkey, [FromServices] MetaGeneratorService metaGenerator) =>
-{
-    var file = await File.ReadAllTextAsync($"./wwwroot/index.html");
-    file = file.Replace("<!-- META -->", await metaGenerator.GetOekakiMetaFor(did, rkey));
-    
-    return Results.Text(file, contentType: "text/html");
-});
 
-// The regex ensures we don't accidentally match the favicon...
-app.MapGet(
-    "/{did:regex(^(?!favicon\\.ico$).*$)}",
-    async ([FromRoute] string did, [FromServices] MetaGeneratorService metaGenerator) =>
-    {
-        var file = await File.ReadAllTextAsync($"./wwwroot/index.html");
-        file = file.Replace("<!-- META -->", await metaGenerator.GetProfileMetaFor(did));
-
-        return Results.Text(file, contentType: "text/html");
-    });
-
-app.MapGet("/ap/note.json",
-    async ([FromQuery] string did, [FromQuery] string rkey, [FromServices] ActivityPubRenderer activityPubRenderer) =>
-    {
-        var response = await activityPubRenderer.RenderNoteForOekaki(did, rkey);
-        if (response is null)
-        {
-            return Results.NotFound();
-        }
-
-        return Results.Json(response, contentType: "application/activity+json");
-    });
-
-app.MapGet("/ap/actor.json",
-    async ([FromQuery] string did, [FromServices] ActivityPubRenderer activityPubRenderer) =>
-    {
-        var response = await activityPubRenderer.RenderActorForProfile(did);
-        if (response is null)
-        {
-            return Results.NotFound();
-        }
-
-        return Results.Json(response, contentType: "application/activity+json");
-    });
-
-app.MapGet("/oembed.json",
-    async ([FromQuery] string url, [FromServices] OEmbedRenderer oEmbedRenderer) =>
-    {
-        var uri = new Uri(url);
-        var split = uri.AbsolutePath.Split("/");
-        var response = await oEmbedRenderer.RenderOEmbedForOekaki(split[1], split[3]);
-        if (response is null)
-        {
-            return Results.NotFound();
-        }
-
-        return Results.Json(response, contentType: "application/json+oembed");
-    });
-
-app.MapGet(
-    "/xrpc/_health",
-    () => new
-    {
-        version = "PinkSea.Gateway"
-    });
+app.MapGeneralEndpoints();
+app.MapActivityPubEndpoints();
+app.MapOEmbedEndpoints();
 
 app.MapFallback(async ([FromServices] MetaGeneratorService metaGenerator) =>
 {
