+++ PinkSea.Gateway/Services/MetaGeneratorService.cs
@@ -1,4 +1,3 @@
-using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Options;
 using PinkSea.Gateway.Lexicons;
 using PinkSea.Gateway.Models;
